agrego metodo calcularMontoTotal a Libreria

+++ ejercicio_libreria/src/main/java/ar/edu/untref/aydoo/ArticuloDeLibreria.java
@@ -2,7 +2,7 @@ package ar.edu.untref.aydoo;
 
 public class ArticuloDeLibreria extends Producto {
 
-	public ArticuloDeLibreria(String nombre, int precio) {
+	public ArticuloDeLibreria(String nombre, double precio) {
 		super.setNombre(nombre);
 		super.setPrecio(precio);
 	}
